requirements.txt
alembic==1.0.7
astroid==2.1.0
autopep8==1.4.3
blinker==1.4
Click==7.0
decorator==4.3.2
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
infinity==1.4
intervals==0.8.1
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
pycodestyle==2.4.0
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
python-slugify==2.0.1
six==1.12.0
SQLAlchemy==1.2.17
Unidecode==1.0.23
validators==0.12.4
Werkzeug==0.14.1
wrapt==1.11.1
WTForms==2.2.1
WTForms-Components==0.10.3
